+++ tests/unit/subsystem/test_update_subsystem.c
@@ -32,6 +32,7 @@ Test(subsystem, test_update_subsystem0)
 	struct yall_subsystem *s = get_fake_subsystem("0", NULL);
 
 	update_subsystem(s, yall_crit, yall_console_output, NULL);
+	cr_assert_str_eq(s->name, "0");
 	cr_assert_eq(s->log_level, yall_crit);
 	cr_assert_eq(s->output_type, yall_console_output);
 	cr_assert_eq(s->file.filename, NULL);
@@ -45,6 +46,7 @@ Test(subsystem, test_update_subsystem0)
 	s = get_fake_subsystem("1", NULL);
 
 	update_subsystem(s, yall_emerg, yall_file_output, "1.log");
+	cr_assert_str_eq(s->name, "1");
 	cr_assert_eq(s->log_level, yall_emerg);
 	cr_assert_eq(s->output_type, yall_file_output);
 	cr_assert_eq(strcmp(s->file.filename, "1.log"), 0);